+++ designer/client/src/actions/nk/editNode.ts
@@ -1,9 +1,15 @@
+import { getEdgesForNode } from "../../components/graph/node-modal/node/NodeDetails";
+
+import { replaceNodeData } from "../../components/graph/node-modal/NodeSwitcherUtils";
+import { Scenario } from "../../components/Process/types";
 import HttpService from "../../http/HttpService";
+import { updateAfterNodeDelete } from "../../reducers/graph/utils";
+import { getGraph } from "../../reducers/selectors/graph";
+import { getProcessDefinitionData } from "../../reducers/selectors/processDefinitionData";
 import { Edge, NodeType, ScenarioGraph, ValidationResult } from "../../types";
 import { ThunkAction } from "../reduxTypes";
 import { calculateProcessAfterChange } from "./calculateProcessAfterChange";
 import { clearProcessCounts } from "./displayProcessCounts";
-import { Scenario } from "../../components/Process/types";
 
 export type EditNodeAction = {
     type: "EDIT_NODE";
@@ -37,5 +43,21 @@ export function editNode(scenarioBefore: Scenario, before: NodeType, after: Node
             validationResult: response.data,
             scenarioGraphAfterChange: scenarioGraph,
         });
+    };
+}
+
+export function replaceNode(before: NodeType, after: NodeType): ThunkAction {
+    return async (dispatch, getState) => {
+        const state = getState();
+        const graph = getGraph(state);
+        const { scenario } = before.id === after.id ? graph : updateAfterNodeDelete(graph, after.id);
+        const { nextEdges: outputEdges, nextNode } = replaceNodeData(
+            before,
+            after,
+            getProcessDefinitionData(state),
+            getEdgesForNode(scenario, before),
+        );
+
+        dispatch(editNode(scenario, before, nextNode, outputEdges));
     };
 }