+++ designer/client/src/components/graph/node-modal/customNode.tsx
@@ -5,6 +5,7 @@ import { AggregateParametersList } from "./aggregateParametersList";
 import { DescriptionField } from "./DescriptionField";
 import { FieldType } from "./editors/field/Field";
 import { IdField } from "./IdField";
+import { isAggregate } from "./isAggregate";
 import { NodeField } from "./NodeField";
 import { ParametersList } from "./parametersList";
 
@@ -40,9 +41,8 @@ export function CustomNode({
     );
 
     const ParametersComponent = useMemo(() => {
-        const isAggregate = ["aggregate-session", "aggregate-sliding", "aggregate-tumbling"].includes(node.nodeType);
-        return isAggregate ? AggregateParametersList : ParametersList;
-    }, [node.nodeType]);
+        return isAggregate(node) ? AggregateParametersList : ParametersList;
+    }, [node]);
 
     return (
         <>