+++ designer/client/src/components/graph/types.ts
@@ -1,26 +1,31 @@
-import { ProcessCounts } from "../../reducers/graph";
 import {
+    editNode,
     injectNode,
     Layout,
     layoutChanged,
     nodeAdded,
     nodesConnected,
     nodesDisconnected,
+    replaceNode,
     resetSelection,
     stickyNoteAdded,
     stickyNoteDeleted,
     stickyNoteUpdated,
     toggleSelection,
 } from "../../actions/nk";
-import { Capabilities } from "../../reducers/selectors/other";
-import { Scenario } from "../Process/types";
 import { StickyNote } from "../../common/StickyNote";
+import { ProcessCounts } from "../../reducers/graph";
+import { Capabilities } from "../../reducers/selectors/other";
+import { NodeType } from "../../types";
+import { Scenario } from "../Process/types";
 
 type ScenarioGraphProps = {
     nodesConnected: typeof nodesConnected;
     nodesDisconnected: typeof nodesDisconnected;
     layoutChanged: typeof layoutChanged;
     injectNode: typeof injectNode;
+    editNode: typeof editNode;
+    replaceNode: typeof replaceNode;
     nodeAdded: typeof nodeAdded;
     stickyNoteAdded: typeof stickyNoteAdded;
     stickyNoteUpdated: typeof stickyNoteUpdated;
@@ -42,6 +47,7 @@ type ScenarioGraphProps = {
     isFragment?: false | null;
 
     connectDropTarget;
+    createFragment?: (callback: (node: NodeType) => void) => void;
 };
 
 type FragmentGraphProps = {