+++ designer/client/src/types/node.ts
@@ -1,5 +1,5 @@
-import { ProcessAdditionalFields, ReturnedType } from "./scenarioGraph";
 import { FragmentInputParameter } from "../components/graph/node-modal/fragment-input-definition/item";
+import { ProcessAdditionalFields, ReturnedType } from "./scenarioGraph";
 import { StickyNoteType } from "./stickyNote";
 
 type Type = "FragmentInput" | typeof StickyNoteType | string;
@@ -18,8 +18,9 @@ export type NodeType<F extends Field = Field> = {
     type: Type;
     isDisabled?: boolean;
     additionalFields?: {
-        description: string;
+        description?: string;
         layoutData?: LayoutData;
+        creatorType?: string;
     };
     parameters?: Parameter[];
     branchParameters?: BranchParams[];